Support insert-final-newline
This commit adds a function applyOnSaveTransformations that is intended to
run the required checks whenever a document is saved. For now it runs the
other new feature of this commit; the insertFinalNewlineTransform, which
checks whether the .editorconfig has 'insert_final_newline: true'. If
insert_final_newline is true, we add a newline if it doesn't already exist.

Signed-off-by: Tor Arvid Lund <torarvid@gmail.com>